created frontend for autofill feature, fixed routing stuff

+++ database/init-db.sql
@@ -14,7 +14,7 @@ CREATE TABLE mangas (
     title VARCHAR(255) NOT NULL,
     status VARCHAR(255) NOT NULL,
     image_url TEXT,
-    max_chapters NUMERIC(10, 1) NOT NULL,
+    max_chapters INTEGER NOT NULL,
     tracking BOOLEAN NOT NULL,
     mangadex_id TEXT,
     last_checked TIMESTAMP NOT NULL
